+++ client/src/App.js
@@ -1,12 +1,12 @@
-import React from 'react'
+import React from "react";
 import {
-	BrowserRouter as Router,
-	Route,
-	Link,
-	Redirect,
-	Switch,
-	withRouter
-} from 'react-router-dom';
+  BrowserRouter as Router,
+  Route,
+  Link,
+  Redirect,
+  Switch,
+  withRouter
+} from "react-router-dom";
 import Auth from "./utils/Auth";
 import API from "./utils/API";
 import Nav from "./components/Nav";
@@ -16,23 +16,27 @@ import { Container } from "./components/Grid";
 import PublicRoute from "./pages/PublicRoute";
 import ProtectedRoute from "./pages/PublicRoute";
 import Title from "./components/Title/title";
-import './App.css';
+import "./App.css";
+import LandingButtons from "./components/LandingButtons/landingButtons";
 
 //I want to add some basic inline styling here, even though we are bringing in styles
 const listStyle = {
-	color: 'cornflowerblue',
-	listStyle: 'none'
+  color: "cornflowerblue",
+  listStyle: "none"
 };
+
 //Now we have all the stuff we need .. let's render some components with the Router
 const AuthExample = () => (
-	<Router>
-		<div className="main-body">
-			<Nav className="App-header" />
-			<Container>
-				<Title />
-				
-				{/* <AuthButton /> */}
-				{/* <ul style={listStyle}>
+  <Router>
+    <div className="main-body">
+      <Nav className="App-header" />
+      <Container>
+        <Title />
+		{/* I would like to link these in the public route, but it won't 
+		let me - If I leave them here, they appear on every page. */}
+        <LandingButtons />
+        {/* <AuthButton /> */}
+        {/* <ul style={listStyle}>
 					<li><Link to="/public">Public Page</Link></li>
 					<li><Link to="/protected">Protected Page</Link></li>
 					<li><Link to="/register">Register a New User</Link></li>
@@ -43,57 +47,54 @@ const AuthExample = () => (
 						Test Route!
 					</button>
 				</ul> */}
-				<Switch>
-					<Route path="/public" component={PublicRoute} />
-					<Route path="/login" component={Login} />
-					<Route path="/register" component={Register} />
-					<PrivateRoute path="/protected" component={ProtectedRoute} />
-					{/* <Route component={NoMatch} /> */}
-				</Switch>
-			</Container>
-		</div>
-	</Router>
-)
-
+        <Switch>
+          <Route path="/public" component={PublicRoute} />
+          <Route path="/login" component={Login} />
+          <Route path="/register" component={Register} />
+          <PrivateRoute path="/protected" component={ProtectedRoute} />
+          {/* <Route component={NoMatch} /> */}
+        </Switch>
+      </Container>
+    </div>
+  </Router>
+);
 
 //Authbutton component / withRouter is imported from react-router
-const AuthButton = withRouter(({ history }) => (
-	Auth.isAuthenticated ? (
-		<div className="container">
-			<p>Success! You are Logged In!</p>
-			<button className="btn btn-danger"
-				onClick={() => {
-					Auth.signout(() => history.push('/'))
-				}}>
-				Sign out
-			</button>
-		</div>
-	) : (
-			<p>You are not logged in.</p>
-		)
-))
+const AuthButton = withRouter(({ history }) =>
+  Auth.isAuthenticated ? (
+    <div className="container">
+      <p>Success! You are Logged In!</p>
+      <button
+        className="btn btn-danger"
+        onClick={() => {
+          Auth.signout(() => history.push("/"));
+        }}
+      >
+        Sign out
+      </button>
+    </div>
+  ) : (
+    <p>You are not logged in.</p>
+  )
+);
 
 // This is the private route component this checks for an authorized user here
 const PrivateRoute = ({ component: Component, ...rest }) => (
-	<Route {...rest} render={props => (
-		Auth.isAuthenticated ? (
-			<Component {...props} />
-		) : (
-				<Redirect to={{
-					pathname: '/login',
-					state: { from: props.location }
-				}} />
-			)
-	)} />
-)
+  <Route
+    {...rest}
+    render={props =>
+      Auth.isAuthenticated ? (
+        <Component {...props} />
+      ) : (
+        <Redirect
+          to={{
+            pathname: "/login",
+            state: { from: props.location }
+          }}
+        />
+      )
+    }
+  />
+);
 
-
-
-
-
-
-
-
-export default AuthExample; 
-
-
+export default AuthExample;
